general cleanup of training code
