+++ mamba_jax/modelling/equinox/loader.py
@@ -37,13 +37,11 @@ def pt_to_raw_pytree(sd, dtype: Optional[jnp.dtype] = jnp.float32):
 
 def init_mamba_from_raw_pytree(tree, config):
     # TODO: add fused add norm and norm type options
-    N = config["d_model"]
-    num_layers = config["n_layer"]
-    vocab_size = config["vocab_size"]
+    # TODO: double check not missing any args
     model = MambaLLM(
-        N,
-        num_layers,
-        vocab_size,
+        config["d_model"],
+        config["n_layer"],
+        config["vocab_size"],
         res_dtype=jnp.float32 if config["residual_in_fp32"] else jnp.bfloat16,
         pad_vocab_mult=config["pad_vocab_size_multiple"],
         kernel_mode=config["kernel_mode"],
